requirements.txt
Django==3.0.4
djangorestframework==3.11.0

flake8>=3.6.0,<3.7.0
